Integrated addCompany and editCompany to one form. Handled datebase.
