+++ src/components/UI/Button/Button.module.css
@@ -1,0 +1,4 @@
+.button {
+  min-width: 100px;
+  height: 45px;
+}
